backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2

requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4

torch==2.0.1
sentence-transformers==2.2.2
huggingface-hub==0.12.1

faiss-cpu==1.7.4

numpy==1.24.4
pandas==1.5.3
scikit-learn==1.3.2
